extended_sxl_specification.md: Switch from using semicolon to colon
Switch from using semicolon ";" to separate blocks, e.g. day of week "d-t;d-t"
to using colon, e.g. "d-t:d-t". Semicolon interferes with the CSV-format which
are used when exporting the SXL to the RSMP simulators. Also <>-chars are
affected (which were used in other commands). This resulted in the RSMP read
the CSV files incorrectly.