S0018: Deprecate "number of time plans"
The S0018 contains the number of time plans in the controller, for
instance "10". But that doesn't mean that time plans 1 - 10 are possible
to use. Instead of knowing the total number of time plans, it is better
to have list of configured time plans that are usable. This is added
by S0022.

Discussed in #72