M0022: Formatting: Use different method for newlines
Instead of using '|' in tables in order to preserve newlines,
use '|br|' instead.

The '|' creates a new paragraph (<p>) which creates unneeded space before
and after. It looks particularly bad in the PDF version.

The '|br|' creates a <br> (HTML) or \\newline (LaTeX) which is a simple
line break, nothing else.